pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "aguasgaia"
authors = [{ name = "Jose Rolo", email = "ze.p.rolo@gmail.com" }]
version = "0.0.17"
description = "Python library to retrieve information from \"Aguas de Gaia\" portal"
readme = "README.md"
requires-python = ">=3.9"
keywords = ["api", "aguasgaia"]
license = { text = "Apache-2.0" }
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "aiohttp",
    "python-dateutil"
]